Get rid of "empty" flag in rss_feed
As far as I can see, it was introduced back in
a64e2296ffd1e62ccb6d26415435471f4fc07856 to indicate when the feed
didn't parse or was empty. But now we have an explicit check for that in
controller, and I couldn't dream up a way to make this field affect
Newsboat's behaviour. Thus I'm removing the field altogether.